Fixed module declarations for tests
