+++ tests/endpoint/delete_endpoint_tests.rs
@@ -1,13 +1,13 @@
 
 #[cfg(test)]
 mod delete_endpoint_tests {
-    use crate::common::test_helpers::{clear_config, get_config2, setup2, EndpointBuilder, HtrsConfigBuilder, ServiceBuilder};
+    use crate::common::test_helpers::{clear_config, get_config, setup, EndpointBuilder, HtrsConfigBuilder, ServiceBuilder};
     use assert_cmd::Command;
     use std::error::Error;
 
     #[test]
     fn given_delete_endpoint_command_with_unknown_service_then_should_fail() -> Result<(), Box<dyn Error>> {
-        let path = setup2(None);
+        let path = setup(None);
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -32,7 +32,7 @@ mod delete_endpoint_tests {
                     .with_name("foo_service")
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -62,7 +62,7 @@ mod delete_endpoint_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -74,7 +74,7 @@ mod delete_endpoint_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.endpoints.len(), 0);
