+++ tests/endpoint/edit_endpoint_tests.rs
@@ -1,13 +1,13 @@
 
 #[cfg(test)]
 mod edit_endpoint_tests {
-    use crate::common::test_helpers::{clear_config, get_config2, setup2, EndpointBuilder, HtrsConfigBuilder, ServiceBuilder};
+    use crate::common::test_helpers::{clear_config, get_config, setup, EndpointBuilder, HtrsConfigBuilder, ServiceBuilder};
     use assert_cmd::Command;
     use std::error::Error;
 
     #[test]
     fn given_edit_endpoint_command_with_unknown_service_then_should_fail() -> Result<(), Box<dyn Error>> {
-        let path = setup2(None);
+        let path = setup(None);
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -32,7 +32,7 @@ mod edit_endpoint_tests {
                     .with_name("foo_service")
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -62,7 +62,7 @@ mod edit_endpoint_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -76,7 +76,7 @@ mod edit_endpoint_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.endpoints.len(), 1);
         let endpoint = &service.endpoints[0];
@@ -101,7 +101,7 @@ mod edit_endpoint_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -115,7 +115,7 @@ mod edit_endpoint_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.endpoints.len(), 1);
         let endpoint = &service.endpoints[0];
@@ -141,7 +141,7 @@ mod edit_endpoint_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -155,7 +155,7 @@ mod edit_endpoint_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.endpoints.len(), 1);
         let endpoint = &service.endpoints[0];
@@ -185,7 +185,7 @@ mod edit_endpoint_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -199,7 +199,7 @@ mod edit_endpoint_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.endpoints.len(), 1);
         let endpoint = &service.endpoints[0];
@@ -230,7 +230,7 @@ mod edit_endpoint_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -244,7 +244,7 @@ mod edit_endpoint_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.endpoints.len(), 1);
         let endpoint = &service.endpoints[0];
