+++ tests/endpoint/list_endpoint_tests.rs
@@ -1,13 +1,13 @@
 
 #[cfg(test)]
 mod list_endpoint_tests {
-    use crate::common::test_helpers::{clear_config, setup, setup2, EndpointBuilder, HtrsConfigBuilder, ServiceBuilder};
+    use crate::common::test_helpers::{clear_config, setup, EndpointBuilder, HtrsConfigBuilder, ServiceBuilder};
     use assert_cmd::Command;
     use std::error::Error;
 
     #[test]
     fn given_list_endpoint_command_with_unknown_service_then_should_fail() -> Result<(), Box<dyn Error>>{
-        let path = setup2(None);
+        let path = setup(None);
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -31,7 +31,7 @@ mod list_endpoint_tests {
                     .with_name("foo_service")
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -65,7 +65,7 @@ mod list_endpoint_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -99,7 +99,7 @@ mod list_endpoint_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
