+++ tests/endpoint/mod.rs
@@ -1,0 +1,4 @@
+mod create_new_endpoint_tests;
+mod delete_endpoint_tests;
+mod edit_endpoint_tests;
+mod list_endpoint_tests;
