+++ tests/environment/create_new_environment_tests.rs
@@ -1,12 +1,12 @@
 #[cfg(test)]
 mod create_new_environment_tests {
-    use crate::common::test_helpers::{clear_config, get_config2, setup2, EnvironmentBuilder, HtrsConfigBuilder, ServiceBuilder};
+    use crate::common::test_helpers::{clear_config, get_config, setup, EnvironmentBuilder, HtrsConfigBuilder, ServiceBuilder};
     use assert_cmd::Command;
     use std::error::Error;
 
     #[test]
     fn given_create_environment_command_without_args_then_should_fail() -> Result<(), Box<dyn Error>> {
-        let path = setup2(None);
+        let path = setup(None);
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -21,7 +21,7 @@ mod create_new_environment_tests {
 
     #[test]
     fn given_create_environment_command_without_service_arg_then_should_fail() -> Result<(), Box<dyn Error>> {
-        let path = setup2(None);
+        let path = setup(None);
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -38,7 +38,7 @@ mod create_new_environment_tests {
 
     #[test]
     fn given_create_environment_command_with_unknown_service_then_should_fail() -> Result<(), Box<dyn Error>> {
-        let path = setup2(None);
+        let path = setup(None);
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -64,7 +64,7 @@ mod create_new_environment_tests {
                     .with_name("foo_service")
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -77,7 +77,7 @@ mod create_new_environment_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.environments.len(), 1);
         let environment = &service.environments[0];
@@ -98,7 +98,7 @@ mod create_new_environment_tests {
                     .with_name("foo_service")
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -113,7 +113,7 @@ mod create_new_environment_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.environments.len(), 1);
         let environment = &service.environments[0];
@@ -134,7 +134,7 @@ mod create_new_environment_tests {
                     .with_name("foo_service")
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -148,7 +148,7 @@ mod create_new_environment_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.environments.len(), 1);
         let environment = &service.environments[0];
@@ -175,7 +175,7 @@ mod create_new_environment_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -209,7 +209,7 @@ mod create_new_environment_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -244,7 +244,7 @@ mod create_new_environment_tests {
                     )
             )
             .build();
-        let path = setup2(Some(config));
+        let path = setup(Some(config));
 
         Command::cargo_bin("htrs")?
             .env("HTRS_CONFIG_PATH", &path)
@@ -258,7 +258,7 @@ mod create_new_environment_tests {
             .assert()
             .success();
 
-        let config = get_config2(&path);
+        let config = get_config(&path);
         let service = &config.services[0];
         assert_eq!(service.environments.len(), 2);
         let existing_environment = &service.environments[0];
