+++ tests/mod.rs
@@ -1,0 +1,4 @@
+mod environment;
+mod endpoint;
+mod common;
+mod service;
