+++ tests/service/mod.rs
@@ -1,0 +1,4 @@
+mod create_new_service_tests;
+mod delete_service_tests;
+mod edit_service_tests;
+mod list_services_tests;
